backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
redis>=5.0.1
httpx>=0.26.0
openai>=1.12.0
google-generativeai>=0.4.0
pandas>=2.2.0
pyarrow>=15.0.0
python-dotenv>=1.0.1
pillow>=11.0.0
greenlet>=3.0.0
